create channel part commited

+++ BackEnd/config/cloudinary.js
@@ -17,7 +17,10 @@ try {
     return result.secure_url
 } catch(error){
     console.log(error)
-    fs.unlinkSync(filePath)
+    if(filePath && fs.existsSync(filePath)){
+        fs.unlinkSync(filePath)
+    }
+    return null;
 }
 }
